Title: Dropdown now works
Description: Dropdown menu can now drop down, with help from ui-bootstrap. ui-bootstrap added as dependency in controller.js, otherwise it wouldn't work. Aperently a dependency to ui-bootstrap is the only thing needed to fix this issue.
Alert: In home.html, the script src for ui-bootstrap is a hyperlink

Commit: Thomas