Title: Local Storage, enviroment modes enabled
Description: Ace editor did not cache your current article before.
Now it caches it in the local storage, and first removes it when, you
create your article. It's persistant over sessions.

I have added an enviroment mode, so the server connects
to a different database, when you run in development or production

important: Still no tests on the client

Commit: Simon