Title: Hotifx on save, and add image
Description: I have removed a bug where and empty image was inserted, when
you cancelled the action in the project creator.

We now have a button, opening the file Uploader on a new tab.

I fixed the toastr in the save project, when there was en error an exception was
thrown, when an error occured

Commit: Simon